+++ interfaces.go
@@ -5,4 +5,5 @@ type IWahoo interface {
 	GetAccessToken(code, uniqueCode string) (*TokenResponse, *RequestError)
 	RefreshToken(refreshToken, uniqueCode string) (*TokenResponse, *RequestError)
 	GetAllWorkout(token string, page int, limit int) (*WorkoutsResponse, *RequestError)
+	DeAuthorize(token string) *RequestError
 }